Divide up the samples into separate files and make the region tags distinctive. Also add graph-specific tests.
